Move PageSize to Export/common
